Digit Recognizer
https://www.kaggle.com/c/digit-recognizer

The solutions to this project are based on various public kernels available in kaggle. For example:
https://www.kaggle.com/cnjn22/digit-recognizer
https://www.kaggle.com/russwill/build-your-own-neural-network-in-r

This project is meant as practice of tools for data science in R. Specifically the use of neural networks using different kinds of packages such as nnet or mxnet (https://mxnet.incubator.apache.org/tutorials/r/mnistCompetition.html).